add sub-sub-directory support for ignored_folder
This commit also "fixes" a subtle "bug" with the previous implementation
of isIgnoredFolder. The previous implementation tested whether a folder
is ignored by (1) splitting the path to be tested using strings.Split with
the seperator "/" and (2) testing whether the first part of the path exists in
w.IgnoredFolders. However, this does not work on architectures that use
a different path seperator. In practice, this "bug" does not present an
issue as an ignored sub-directory is skipped entirely by filepath.Walk.